Refactor tests: Break apart tests/test_main.py into two files - test_generate_headline
